Address code review feedback
- Tighten Phase 1 test: check g_dtor_calls == 1 (not >=1) to catch
  double-destructor bugs
- Log a warning when emitInlineDestructorCall cannot find a variable
  instead of silently skipping it
- Use size_t for capture_try_cleanup_depth_ to match scope_stack_.size()
  type, removing the signed/unsigned cast

+++ src/CodeGen_Helpers.cpp
@@ -6,7 +6,7 @@ void AstToIr::exitScope() {
 	if (!scope_stack_.empty()) {
 		// If try-cleanup capture is active and this is the target scope depth,
 		// record vars in LIFO order before emitting their destructors.
-		if (capture_try_cleanup_ && (int)scope_stack_.size() == capture_try_cleanup_depth_) {
+		if (capture_try_cleanup_ && scope_stack_.size() == capture_try_cleanup_depth_) {
 			for (auto it = scope_stack_.back().rbegin(); it != scope_stack_.back().rend(); ++it) {
 				captured_try_cleanup_vars_.push_back(*it);
 			}
